added OTS to telemetry overlay for Super Formula cars
ugly for now but fully functional. shows on/off state of OTS, remaining time left during the race, and a circular countdown timer that lets you know when you can use OTS again, along with very clear (but ugly for now) color changes depending on the state. Green is enabled, purple is disabled but cooling down, and blue is disabled and ready to use.